[Dy2St] Move more transformer specific utilities to transformers dir
