+++ python/paddle/jit/dy2static/transformers/loop_transformer.py
@@ -19,24 +19,28 @@ from paddle.base import unique_name
 from paddle.utils import gast
 
 from ..utils import (
-    FOR_BODY_PREFIX,
-    FOR_CONDITION_PREFIX,
-    WHILE_BODY_PREFIX,
-    WHILE_CONDITION_PREFIX,
-    FunctionNameLivenessAnalysis,
     GetterSetterHelper,
     ast_to_source_code,
-    create_get_args_node,
-    create_nonlocal_stmt_nodes,
-    create_set_args_node,
 )
 from .base import (
     BaseTransformer,
     ForLoopTuplePreTransformer,
     ForNodeVisitor,
 )
-from .ifelse_transformer import ARGS_NAME
-from .utils import create_name_str, get_attribute_full_name, get_parent_mapping
+from .utils import (
+    ARGS_NAME,
+    FOR_BODY_PREFIX,
+    FOR_CONDITION_PREFIX,
+    WHILE_BODY_PREFIX,
+    WHILE_CONDITION_PREFIX,
+    FunctionNameLivenessAnalysis,
+    create_get_args_node,
+    create_name_str,
+    create_nonlocal_stmt_nodes,
+    create_set_args_node,
+    get_attribute_full_name,
+    get_parent_mapping,
+)
 
 __all__ = []
